feat(tests): add mock_metadata fixture for unit tests
