+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import datetime
 import io
 from pathlib import Path
 from unittest.mock import patch
@@ -5,6 +6,8 @@ from unittest.mock import patch
 import numpy as np
 import pytest
 from PIL import Image
+
+from focoos.ports import FocoosTask, ModelMetadata, ModelStatus
 
 
 @pytest.fixture
@@ -42,3 +45,26 @@ def image_path(tmp_path: Path, pil_image: Image.Image) -> Path:
 def image_ndarray(pil_image: Image.Image) -> np.ndarray:
     """Convert PIL image to NumPy array (which is in RGB format)."""
     return np.array(pil_image)
+
+
+@pytest.fixture
+def mock_metadata():
+    return ModelMetadata(
+        ref="test_model_ref",
+        name="Test Model",
+        description="A test model for unit tests",
+        owner_ref="test_owner",
+        focoos_model="test_focoos_model",
+        task=FocoosTask.DETECTION,
+        created_at=datetime.datetime.now(),
+        updated_at=datetime.datetime.now(),
+        status=ModelStatus.DEPLOYED,
+        metrics={"accuracy": 0.95},
+        latencies=[{"inference": 0.1}],
+        classes=["class_0", "class_1"],
+        im_size=640,
+        hyperparameters=None,
+        training_info=None,
+        location=None,
+        dataset=None,
+    )
